Sort old sets to MCI order
Older core sets and expansion before Exodus didn't have card numbers.
MCI sorts most of these sets in BUGRWALX order but a few are in slightly
different orders.

To solve this, we need to sort the cards in each of these sets in the
appropriate MCI order and then set their number in that order before further processing.